Refactoring to remove dupe code
